mcpserver/pyproject.toml
[project]
name = "mcpserver"
version = "0.1.0"
description = "Custom LLM Chat App with MCP Filesystem Server"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.14.1",
    "python-dotenv>=1.0.0",
    "langchain-google-genai>=1.0.0",
    "langchain>=0.1.0",
    "langchain-core>=0.1.0",
    "streamlit>=1.28.0",
    "google-generativeai>=0.3.0",
    "asyncio-mqtt>=0.16.0",
    "nest-asyncio>=1.5.0", 
]
